Remove `JdbcConfig` and rework visibility handling in Artid entity
Eliminated custom JDBC converters for `VISIBILITY_STATE` by removing `JdbcConfig`. Refactored `Artid` entity to handle `visibilityState` as a `String` instead of the enum type. Updated database queries and models to align with the new visibility state handling design. Adjusted DTOs and services to support these changes, ensuring consistency across the codebase.

+++ app/artid-server/src/main/java/afam/artidserver/model/dto/ExternalShareArtIDResponse.java
@@ -5,7 +5,7 @@ import java.time.OffsetDateTime;
 public record ExternalShareArtIDResponse(
         Long id,
         Long idArtid,
-        Long clickCounter,
+        Integer clickCounter,
         Boolean isActive,
         OffsetDateTime expirationDate,
         OffsetDateTime lastOpened,
