+++ app/artid-server/src/main/java/afam/artidserver/model/dto/ExternalShareResponse.java
@@ -8,7 +8,7 @@ import java.time.OffsetDateTime;
 public record ExternalShareResponse (
     Long id,
     Long idArtid,
-    Long clickCounter,
+    Integer clickCounter,
     Boolean isActive,
     OffsetDateTime expirationDate,
     OffsetDateTime lastOpened
